add random match custom data send func

+++ src/game/systems/matchs/random_match_leave_system.rs
@@ -6,11 +6,11 @@ use crate::{
     game::{
         components::matchs::{random_match_leave_component::RandomMatchLeaveSendTo}, 
         memory::user::user_memory::get_user_socket, 
-        enums::core_enum::MessageType
+        enums::core_enum::MessageType, systems::message::system_message_system::system_message_send
     },
     database::redis::{
         matchs::match_hash::{
-            delete_my_match, get_match_members, get_my_match, delete_match_member,
+            delete_my_match, get_match_members, get_my_match, delete_match_member, check_my_match_exist,
         }, 
         socket::socket_hash::get_my_info
     },
@@ -49,9 +49,13 @@ pub fn random_match_leave(
     let sender_info = get_my_info(&send_uid).unwrap();
     let username = sender_info.iter().find(|(key, _)| *key == "name").map(|(_, value)| value.to_owned()).unwrap();
     let match_id = get_my_match(send_uid).unwrap();
+    if !check_my_match_exist(&send_uid).unwrap() {
+        system_message_send(&send_uid, format!("you are not in random_match."));
+        return;
+    }
+    let match_user_list = get_match_members(&match_id).unwrap();
     delete_match_member(&match_id, send_uid).unwrap();
     delete_my_match(send_uid).unwrap();
-    let match_user_list = get_match_members(&match_id).unwrap();
     let sockets = get_user_socket();
     let user_sockets = sockets.iter().filter(|user_sockets| match_user_list.contains(user_sockets.0)).map(|(_, user_socket)| user_socket);
